Debayer in top-down orientation by default now. If the new ROWORDER keyword is found we follow this keyword (BOTTOM-UP or TOP-DOWN).

+++ src/algos/photometry.c
@@ -287,7 +287,7 @@ photometry *getPhotometryData(gsl_matrix* z, fitted_PSF *psf, gboolean verbose) 
 	r2 *= r2;
 	rmin_sq = (appRadius - 0.5) * (appRadius - 0.5);
 
-	size_t ndata = (y2 - y1) * (x2 - x1);
+	int ndata = (y2 - y1) * (x2 - x1);
 	if (ndata <= 0) {
 		siril_log_color_message(_("An error occurred in your selection. Please make another selection.\n"), "red");
 		return NULL;
